refactor: レシピ詳細をコンポーネントに分割
