+++ src/app/(authenticated)/recipes/[recipeId]/page.tsx
@@ -1,82 +1,50 @@
 'use client';
 
-import { stepsItemForParse } from '@/lib/parser/stepsItemForParse';
-import { numberSchema } from '@/lib/validators/numberSchema';
-import { recipeDetailSchema } from '@/lib/validators/recipeSchema';
+import { numberSchema } from '@/lib/schema/numberSchema';
+import { recipeDetailSchema } from '@/lib/schema/recipeSchema';
+import { userRoomSchema } from '@auth/lib/validation/userRoomSchema';
+import { Loading } from '@authenticated/components/Loading';
 import { useAuthedSWR } from '@authenticated/hooks/useAuthedSWR';
+import { useRecipes } from '@authenticated/hooks/useRecipes';
+import { RecipeItem } from '@authenticated/recipes/components/RecipeItem';
 import Link from 'next/link';
-import { notFound, useParams, useSearchParams } from 'next/navigation';
+import { useParams } from 'next/navigation';
 
 export default function RecipeDetailPage() {
   const params = useParams();
-  const result = numberSchema.safeParse(params.recipeId);
-  if (!result.success) notFound();
-  const recipeId = result.data;
+  const parsedParams = numberSchema.safeParse(params.recipeId);
+  if (!parsedParams.success) return <p>不正なURLです</p>;
 
-  const searchParams = useSearchParams();
-  const from = searchParams.get('from');
+  const recipeId = parsedParams.data;
 
   const {
     data: recipe,
-    error,
-    isLoading,
-  } = useAuthedSWR(`/api/recipes/${recipeId}`, recipeDetailSchema);
+    errorMsg: recipeErrorMsg,
+    isLoading: isRecipsLoading,
+  } = useRecipes(`/api/recipes/${recipeId}`, recipeDetailSchema);
 
-  if (isLoading) return <p>読み込み中...</p>;
-  if (error) return <p>エラー: {String(error)}</p>;
-  if (!recipe) return <p>レシピがありません</p>;
+  const { data: userRoom } = useAuthedSWR('/api/userRoom', userRoomSchema);
+  if (!userRoom) return <Loading />;
 
-  const itemsBlock = stepsItemForParse(recipe.ingredients, recipe.instructions);
+  const talkRoomId = userRoom.talkRoom.id;
+
+  if (recipeErrorMsg) {
+    return <p>{recipeErrorMsg}</p>;
+  }
+  if (!recipe) {
+    return <Loading />;
+  }
 
   return (
     <main className="p-6">
       <h1 className="text-2xl font-bold mb-2">{recipe.title}</h1>
-      {recipe.imageKey && (
-        <img
-          src={`/images/${recipe.imageKey}`}
-          alt={recipe.title}
-          className="mb-4 rounded-lg shadow"
-        />
-      )}
-      <p className="mb-2 text-gray-600">
-        ⏱ 調理時間: {recipe.cookingTime || '不明'}
-      </p>
-      <p className="mb-4">{recipe.point}</p>
-      <section>
-        <h2 className="text-lg font-semibold mb-1">材料（2人分）</h2>
-        <ul className="list-disc list-inside mb-4">
-          {itemsBlock.ingredients.map((item: string, i: number) => (
-            <li key={i}>{item}</li>
-          ))}
-        </ul>
-
-        <h2 className="text-lg font-semibold mb-1">作り方</h2>
-        <ol className="list-decimal list-inside">
-          {itemsBlock.instructions.map((step: string, i: number) => (
-            <li key={i} className="mb-1">
-              {step.replace(/^\d+[:：]\s*/, '').trim()}
-            </li>
-          ))}
-        </ol>
-      </section>
-      {recipe && (
-        <Link
-          href={`/recipes?from=${from}`}
-          prefetch={false}
-          className="text-sm underline"
-        >
-          すべてのレシピを見る
-        </Link>
-      )}
-      {from && (
-        <Link
-          href={`/talkRoom/${from}`}
-          prefetch={false}
-          className="text-sm underline"
-        >
-          会話に戻る
-        </Link>
-      )}
+      <RecipeItem recipe={recipe} />
+      <Link href={`/recipes`} className="text-sm underline">
+        すべてのレシピを見る
+      </Link>
+      <Link href={`/talkRoom/${talkRoomId}`} className="text-sm underline">
+        会話に戻る
+      </Link>
     </main>
   );
 }
