+++ src/app/(authenticated)/recipes/_components/RecipeItem.tsx
@@ -1,0 +1,43 @@
+import { stepsItemForParse } from '@/lib/parser/stepsItemForParse';
+import { RecipeDetail } from '@/lib/schema/recipeSchema';
+
+type Props = {
+  recipe: RecipeDetail;
+};
+
+export const RecipeItem = ({ recipe }: Props) => {
+  const itemsBlock = stepsItemForParse(recipe.ingredients, recipe.instructions);
+
+  return (
+    <>
+      {recipe.imageKey && (
+        <img
+          src={`/images/${recipe.imageKey}`}
+          alt={recipe.title}
+          className="mb-4 rounded-lg shadow"
+        />
+      )}
+      <p className="mb-2 text-gray-600">
+        ⏱ 調理時間: {recipe.cookingTime || '不明'}
+      </p>
+      <p className="mb-4">{recipe.point}</p>
+      <section>
+        <h2 className="text-lg font-semibold mb-1">材料（2人分）</h2>
+        <ul className="list-disc list-inside mb-4">
+          {itemsBlock.ingredients.map((item: string, i: number) => (
+            <li key={i}>{item}</li>
+          ))}
+        </ul>
+
+        <h2 className="text-lg font-semibold mb-1">作り方</h2>
+        <ol className="list-decimal list-inside">
+          {itemsBlock.instructions.map((step: string, i: number) => (
+            <li key={i} className="mb-1">
+              {step.replace(/^\d+[:：]\s*/, '').trim()}
+            </li>
+          ))}
+        </ol>
+      </section>
+    </>
+  );
+};
